% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/befehle.R
\name{bemerkung}
\alias{bemerkung}
\title{Befehle für FileMaker-kompatible Aufbereitung}
\usage{
bemerkung(data, a, b)
}
\arguments{
\item{data}{ein Ausgangsdatensatz.}

\item{a}{eine Variable}

\item{b}{eine Variable}
}
\value{
Erzeugen neuer Variablen in \code{data} aus \code{a}, \code{b}, etc.
}
\description{
Befehle für FileMaker-kompatible Aufbereitung
}
