% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail.R
\name{gmail}
\alias{gmail}
\title{Erzeugen der Variable Adr_e.mail}
\usage{
gmail(data, a)
}
\arguments{
\item{data}{der Ausgangsdatensatz}

\item{a}{erste Variable}
}
\value{
Die Variable \code{a} wird zu Adr_e.Mail umbenannt und alle Google-Mail-Adressen werden zu gmail.com standardisiert
}
\description{
Erzeugen der Variable Adr_e.mail
}
