% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selektion17.R
\name{selektion17}
\alias{selektion17}
\title{Erzeugen der Variable Selektion_17}
\usage{
selektion17(data, a)
}
\arguments{
\item{data}{der Ausgangsdatensatz}

\item{a}{erste Variable}
}
\value{
Erzeugen der Variable Selektion_17 aus \code{a}
}
\description{
Erzeugen der Variable Selektion_17
}
