% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitschrift.R
\name{zeitschrift}
\alias{zeitschrift}
\title{Erzeugen der Variable Mitgl_digitale_Zeitschrift}
\usage{
zeitschrift(data, a)
}
\arguments{
\item{data}{ein Ausgangsdatensatz.}

\item{a}{erste Variable}
}
\value{
Erzeugen der Variable Mitgl_digitale_Zeitschrift aus \code{a}.
}
\description{
Erzeugen der Variable Mitgl_digitale_Zeitschrift
}
