% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemerkung.R
\name{bemerkung}
\alias{bemerkung}
\title{Erzeugen der Variable Sel_Bemerkung}
\usage{
bemerkung(data, a, b)
}
\arguments{
\item{data}{ein Ausgangsdatensatz}

\item{a}{erste Variable}

\item{b}{zweite Variable}
}
\value{
Erzeugen der Variable Sel_Bemerkung aus \code{a}, \code{b}.
}
\description{
Erzeugen der Variable Sel_Bemerkung
}
