% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail.R
\name{gmail}
\alias{gmail}
\title{Erzeugen der Variable gmail}
\usage{
gmail(data, a)
}
\arguments{
\item{data}{der Ausgangsdatensatz}

\item{a}{erste Variable}
}
\value{
Alle Google-Mail Adressen in der Variable \code{a} werden zu gmail.com standardisiert
}
\description{
Erzeugen der Variable gmail
}
